+++ src/compat.jl
@@ -1,3 +1,19 @@
+"""
+    @public args...
+
+declares `args...` as being `public` even though they are not exported. For Julia version
+< 1.11, this macro does nothing. Using this macro also avoid errors with CI and coverage
+tools.
+
+"""
+macro public(args::Union{Symbol,Expr}...)
+    if VERSION ≥ v"1.11.0-DEV.469"
+        esc(Expr(:public, args...))
+    else
+        nothing
+    end
+end
+
 if VERSION < v"1.7.0-beta1"
     # `Returns` is not defined prior to Julia 1.7.
     """